+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+## 1.0.5
+
+* Update [Repositry (Github)](https://github.com/vvvirani/flutter_amazon_payfort)
+
+## 1.0.4
+
+* Improve performance and bug fixes
+
+## 1.0.3
+
+* Improve performance and bug fixes
+
+## 1.0.2
+
+* Solved payment failed issue from `iOS` side(see issue [#13](https://github.com/vvvirani/flutter_amazon_payfort/issues/13)).
+
+## 1.0.1
+
+* Fixed payment issue on release mode from `Android` side
+
+## 1.0.0
+
+* Improve performance and bug fixes
+
+## 0.0.9
+
+* Upgrade Packages version
+* Fix the payment issue in the release mode from the `iOS` side, Currenlty working from the `Android` side
+
 ## 0.0.8
 
 * Fix the `succeeded` call back issue for card payment
@@ -34,5 +63,5 @@
 
 ## 0.0.1
 
-* Initial release.
+* Initial release